requirements.txt
fastapi
uvicorn
jinja2
python-multipart
numpy
chance
Pillow
matplotlib
opencv-python
